third_party/blink/web_tests/platform/generic/virtual/non-overlay-scrollbar/body-quirk-client-size-expected.txt

Style recalc when reading clientWidth/clientHeight of body in quirks mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object Internals] is defined.
PASS internals.overlayScrollbarsEnabled is false
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS 800 is 800
PASS 600 is > 585
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS 800 is > 785
PASS 600 is 600
PASS successfullyParsed is true

TEST COMPLETE
